requirements.txt
amqp==5.0.9
asgiref==3.5.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptocode==0.1
Django==3.2.11
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-cors-headers==3.11.0
django-timezone-field==4.2.3
djangorestframework==3.13.1
idna==3.3
kombu==5.2.3
prompt-toolkit==3.0.24
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pycryptodomex==3.13.0
PyPDF2==1.26.0
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
